+++ apis/cluster/v1alpha1/clustermanager_types.go
@@ -15,6 +15,7 @@ limitations under the License.
 package v1alpha1
 
 import (
+	"os"
 	"strings"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -122,24 +123,25 @@ type ProviderVsphereSpec struct {
 
 // ClusterManagerStatus defines the observed state of ClusterManager
 type ClusterManagerStatus struct {
-	Provider               string              `json:"provider,omitempty"`
-	Version                string              `json:"version,omitempty"`
-	Ready                  bool                `json:"ready,omitempty"`
-	MasterRun              int                 `json:"masterRun,omitempty"`
-	WorkerRun              int                 `json:"workerRun,omitempty"`
-	FailureReason          *string             `json:"failureReason,omitempty"`
-	Phase                  ClusterManagerPhase `json:"phase,omitempty"`
-	ControlPlaneEndpoint   string              `json:"controlPlaneEndpoint,omitempty"`
-	ArgoReady              bool                `json:"argoReady,omitempty"`
-	TraefikReady           bool                `json:"traefikReady,omitempty"`
-	MonitoringReady        bool                `json:"gatewayReady,omitempty"`
-	AuthClientReady        bool                `json:"authClientReady,omitempty"`
-	HyperregistryOidcReady bool                `json:"hyperregistryOidcReady,omitempty"`
-	ApplicationLink        bool                `json:"applicationLink,omitempty"`
-	OpenSearchReady        bool                `json:"openSearchReady,omitempty"`
-	Conditions             []metav1.Condition  `json:"conditions,omitempty"`
+	Provider              string              `json:"provider,omitempty"`
+	Version               string              `json:"version,omitempty"`
+	Ready                 bool                `json:"ready,omitempty"`
+	MasterRun             int                 `json:"masterRun,omitempty"`
+	WorkerRun             int                 `json:"workerRun,omitempty"`
+	FailureReason         *string             `json:"failureReason,omitempty"`
+	Phase                 ClusterManagerPhase `json:"phase,omitempty"`
+	ControlPlaneEndpoint  string              `json:"controlPlaneEndpoint,omitempty"`
+	Conditions            []metav1.Condition  `json:"conditions,omitempty"`
+	ArgoReady             bool                `json:"argoReady,omitempty"`
+	TraefikReady          bool                `json:"traefikReady,omitempty"`
+	GatewayReady          bool                `json:"gatewayReady,omitempty"`
+	GatewayReadyMigration bool                `json:"gatewayReadyMigration,omitempty"`
+	AuthClientReady       bool                `json:"authClientReady,omitempty"`
+	OpenSearchReady       bool                `json:"openSearchReady,omitempty"`
+	ApplicationLink       string              `json:"applicationLink,omitempty"`
 	// will be deprecated
 	PrometheusReady bool `json:"prometheusReady,omitempty"`
+	// HyperregistryOidcReady bool                    `json:"hyperregistryOidcReady,omitempty"`
 }
 
 type ClusterManagerPhase string
@@ -156,6 +158,17 @@ const (
 	ClusterManagerPhaseDeleting = ClusterManagerPhase("Deleting")
 	// 클러스터 생성에 실패한 상태
 	ClusterManagerPhaseFailed = ClusterManagerPhase("Failed")
+)
+
+// deprecated phases
+const (
+	ClusterManagerDeprecatedPhasePending      = ClusterManagerPhase("Pending")
+	ClusterManagerDeprecatedPhaseProvisioning = ClusterManagerPhase("Provisioning")
+	ClusterManagerDeprecatedPhaseRegistering  = ClusterManagerPhase("Registering")
+	ClusterManagerDeprecatedPhaseProvisioned  = ClusterManagerPhase("Provisioned")
+	ClusterManagerDeprecatedPhaseRegistered   = ClusterManagerPhase("Registered")
+	ClusterManagerDeprecatedPhaseFailed       = ClusterManagerPhase("Failed")
+	ClusterManagerDeprecatedPhaseUnknown      = ClusterManagerPhase("Unknown")
 )
 
 const (
@@ -285,4 +298,22 @@ func (c *ClusterManager) GetConditions() []metav1.Condition {
 
 func (c *ClusterManager) SetConditions(conditions []metav1.Condition) {
 	c.Status.Conditions = conditions
+}
+
+func (c *ClusterManager) SetApplicationLink(subdomain string) {
+	c.Status.ApplicationLink = strings.Join(
+		[]string{
+			"https://",
+			subdomain,
+			".",
+			os.Getenv("HC_DOMAIN"),
+			"/applications/",
+			c.GetNamespacedPrefix(),
+			"-applications?node=argoproj.io/Application/argocd/",
+			c.GetNamespacedPrefix(),
+			"-applications/0&resource=",
+			"",
+		},
+		"",
+	)
 }