+++ controllers/hyperAuth/apiCaller.go
@@ -46,8 +46,8 @@ func GetTokenAsAdmin(secret *coreV1.Secret) (string, error) {
 
 	// Make Request Object
 	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_GET_TOKEN, nil)
-	playload := strings.NewReader(data.Encode())
-	req, err := http.NewRequest(http.MethodPost, url, playload)
+	payload := strings.NewReader(data.Encode())
+	req, err := http.NewRequest(http.MethodPost, url, payload)
 	if err != nil {
 		return "", err
 	}
@@ -71,7 +71,7 @@ func GetTokenAsAdmin(secret *coreV1.Secret) (string, error) {
 	}
 	accessToken := resultJson["access_token"].(string)
 
-	return accessToken, nil
+	return strings.Join([]string{"Bearer", accessToken}, " "), nil
 }
 
 func GetIdByClientId(clientId string, secret *coreV1.Secret) (string, error) {
@@ -85,7 +85,7 @@ func GetIdByClientId(clientId string, secret *coreV1.Secret) (string, error) {
 	if err != nil {
 		return "", err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -111,7 +111,7 @@ func GetIdByClientId(clientId string, secret *coreV1.Secret) (string, error) {
 		}
 	}
 
-	return "", nil
+	return "", HyperAuthError{NotFound: true, Type: RESOURCE_TYPE_CLIENT, Name: clientId}
 }
 
 func CreateClient(config ClientConfig, secret *coreV1.Secret) error {
@@ -120,27 +120,19 @@ func CreateClient(config ClientConfig, secret *coreV1.Secret) error {
 		return err
 	}
 
-	id, err := GetIdByClientId(config.ClientId, secret)
-	if err != nil {
-		return err
-	}
-	if IsClientExist(id) {
-		return nil
-	}
-
 	jsonData, err := json.Marshal(config)
 	if err != nil {
 		return err
 	}
-	playload := bytes.NewBuffer(jsonData)
+	payload := bytes.NewBuffer(jsonData)
 
 	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_CREATE_CLIENT, nil)
-	req, err := http.NewRequest(http.MethodPost, url, playload)
+	req, err := http.NewRequest(http.MethodPost, url, payload)
 	if err != nil {
 		return err
 	}
 	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -167,26 +159,23 @@ func CreateClientLevelProtocolMapper(config ClientLevelProtocolMapperConfig, sec
 	if err != nil {
 		return err
 	}
-	if !IsClientExist(id) {
-		return fmt.Errorf("client not found")
-	}
 
 	jsonData, err := json.Marshal(config.ProtocolMapper)
 	if err != nil {
 		return err
 	}
-	playload := bytes.NewBuffer(jsonData)
+	payload := bytes.NewBuffer(jsonData)
 
 	params := map[string]string{
 		"id": id,
 	}
 	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_CREATE_CLIENT_PROTOCOL_MAPPERS, params)
-	req, err := http.NewRequest(http.MethodPost, url, playload)
+	req, err := http.NewRequest(http.MethodPost, url, payload)
 	if err != nil {
 		return err
 	}
 	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -213,9 +202,6 @@ func CreateClientLevelRole(config ClientLevelRoleConfig, secret *coreV1.Secret) 
 	if err != nil {
 		return err
 	}
-	if !IsClientExist(id) {
-		return fmt.Errorf("client not found")
-	}
 
 	data := RoleConfig{
 		Name: config.Role.Name,
@@ -224,18 +210,18 @@ func CreateClientLevelRole(config ClientLevelRoleConfig, secret *coreV1.Secret) 
 	if err != nil {
 		return err
 	}
-	playload := bytes.NewBuffer(jsonData)
+	payload := bytes.NewBuffer(jsonData)
 
 	params := map[string]string{
 		"id": id,
 	}
 	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_CREATE_CLIENT_ROLES, params)
-	req, err := http.NewRequest(http.MethodPost, url, playload)
+	req, err := http.NewRequest(http.MethodPost, url, payload)
 	if err != nil {
 		return err
 	}
 	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -260,14 +246,14 @@ func GetUserIdByEmail(userEmail string, secret *coreV1.Secret) (string, error) {
 
 	params := map[string]string{
 		"userEmail": userEmail,
-		"token":     token,
+		"token":     strings.Replace(token, "Bearer ", "", 1),
 	}
 	url := SetServiceDomainURI(HYPERAUTH_SERVICE_GET_USER_ID_BY_EMAIL, params)
 	req, err := http.NewRequest(http.MethodGet, url, nil)
 	if err != nil {
 		return "", err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -287,7 +273,7 @@ func GetUserIdByEmail(userEmail string, secret *coreV1.Secret) (string, error) {
 		return "", err
 	}
 	if respJson == nil {
-		return "", fmt.Errorf("user not found")
+		return "", HyperAuthError{NotFound: true, Type: RESOURCE_TYPE_USER_EMAIL, Name: userEmail}
 	}
 
 	return respJson.Id, nil
@@ -303,9 +289,6 @@ func GetClientRoleIdByRoleName(clientId string, roleName string, secret *coreV1.
 	if err != nil {
 		return "", err
 	}
-	if !IsClientExist(id) {
-		return "", fmt.Errorf("client not found")
-	}
 
 	params := map[string]string{
 		"id":       id,
@@ -316,7 +299,7 @@ func GetClientRoleIdByRoleName(clientId string, roleName string, secret *coreV1.
 	if err != nil {
 		return "", err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -336,7 +319,7 @@ func GetClientRoleIdByRoleName(clientId string, roleName string, secret *coreV1.
 		return "", err
 	}
 	if respJson == nil {
-		return "", fmt.Errorf("role not found")
+		return "", HyperAuthError{NotFound: true, Type: RESOURCE_TYPE_CLIENT_ROLE, Name: roleName}
 	}
 
 	return respJson.Id, nil
@@ -351,9 +334,6 @@ func AddClientLevelRolesToUserRoleMapping(config ClientLevelRoleConfig, userEmai
 	id, err := GetIdByClientId(config.ClientId, secret)
 	if err != nil {
 		return err
-	}
-	if !IsClientExist(id) {
-		return fmt.Errorf("client not found")
 	}
 
 	userId, err := GetUserIdByEmail(userEmail, secret)
@@ -375,19 +355,19 @@ func AddClientLevelRolesToUserRoleMapping(config ClientLevelRoleConfig, userEmai
 	if err != nil {
 		return err
 	}
-	playload := bytes.NewBuffer(jsonData)
+	payload := bytes.NewBuffer(jsonData)
 
 	params := map[string]string{
 		"userId": userId,
 		"id":     id,
 	}
 	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_ADD_CLIENT_ROLE_TO_USER, params)
-	req, err := http.NewRequest(http.MethodPost, url, playload)
+	req, err := http.NewRequest(http.MethodPost, url, payload)
 	if err != nil {
 		return err
 	}
 	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -418,7 +398,7 @@ func GetRealmRoleIdByRoleName(roleName string, secret *coreV1.Secret) (string, e
 	if err != nil {
 		return "", err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -438,7 +418,7 @@ func GetRealmRoleIdByRoleName(roleName string, secret *coreV1.Secret) (string, e
 		return "", err
 	}
 	if respJson == nil {
-		return "", fmt.Errorf("role not found")
+		return "", HyperAuthError{NotFound: true, Type: RESOURCE_TYPE_REALM_ROLE, Name: roleName}
 	}
 
 	return respJson.Id, nil
@@ -469,18 +449,18 @@ func AddRealmLevelRolesToUserRoleMapping(roleName string, userEmail string, secr
 	if err != nil {
 		return err
 	}
-	playload := bytes.NewBuffer(jsonData)
+	payload := bytes.NewBuffer(jsonData)
 
 	params := map[string]string{
 		"userId": userId,
 	}
 	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_ADD_REALM_ROLE_TO_USER, params)
-	req, err := http.NewRequest(http.MethodPost, url, playload)
+	req, err := http.NewRequest(http.MethodPost, url, payload)
 	if err != nil {
 		return err
 	}
 	req.Header.Add("Content-Type", "application/json")
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -508,7 +488,7 @@ func GetClientScopesIdByName(name string, secret *coreV1.Secret) (string, error)
 	if err != nil {
 		return "", err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -534,7 +514,7 @@ func GetClientScopesIdByName(name string, secret *coreV1.Secret) (string, error)
 		}
 	}
 
-	return "", fmt.Errorf("client scope not found")
+	return "", HyperAuthError{NotFound: true, Type: RESOURCE_TYPE_CLIENT_SCOPE, Name: name}
 }
 
 func AddClientScopeToClient(config ClientScopeMappingConfig, secret *coreV1.Secret) error {
@@ -546,9 +526,6 @@ func AddClientScopeToClient(config ClientScopeMappingConfig, secret *coreV1.Secr
 	id, err := GetIdByClientId(config.ClientId, secret)
 	if err != nil {
 		return err
-	}
-	if !IsClientExist(id) {
-		return nil
 	}
 
 	clientScopeId, err := GetClientScopesIdByName(config.ClientScope.Name, secret)
@@ -565,7 +542,7 @@ func AddClientScopeToClient(config ClientScopeMappingConfig, secret *coreV1.Secr
 	if err != nil {
 		return err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)
@@ -582,6 +559,124 @@ func AddClientScopeToClient(config ClientScopeMappingConfig, secret *coreV1.Secr
 	return nil
 }
 
+func CreateGroup(config GroupConfig, secret *coreV1.Secret) error {
+	token, err := GetTokenAsAdmin(secret)
+	if err != nil {
+		return err
+	}
+
+	jsonData, err := json.Marshal(config)
+	if err != nil {
+		return err
+	}
+	payload := bytes.NewBuffer(jsonData)
+
+	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_CREATE_GROUP, nil)
+	req, err := http.NewRequest(http.MethodPost, url, payload)
+	if err != nil {
+		return err
+	}
+
+	req.Header.Add("Content-Type", "application/json")
+	req.Header.Add("Authorization", token)
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	if !IsOK(resp.StatusCode) {
+		reqDump, _ := httputil.DumpRequest(req, true)
+		return fmt.Errorf("failed to create group: " + string(reqDump) + "\n" + resp.Status)
+	}
+
+	return nil
+}
+
+func GetGroupIdByName(name string, secret *coreV1.Secret) (string, error) {
+	token, err := GetTokenAsAdmin(secret)
+	if err != nil {
+		return "", err
+	}
+
+	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_GET_GROUP, nil)
+	req, err := http.NewRequest(http.MethodGet, url, nil)
+	if err != nil {
+		return "", err
+	}
+	req.Header.Add("Authorization", token)
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return "", err
+	}
+	defer resp.Body.Close()
+
+	reqDump, _ := httputil.DumpRequest(req, true)
+	if !IsOK(resp.StatusCode) {
+		return "", fmt.Errorf("failed to get group: " + string(reqDump) + "\n" + resp.Status)
+	}
+
+	respJson := &[]GroupConfig{}
+	err = json.NewDecoder(resp.Body).Decode(respJson)
+	if err != nil {
+		return "", err
+	}
+
+	for _, data := range *respJson {
+		if data.Name == name {
+			return data.Id, nil
+		}
+	}
+
+	return "", HyperAuthError{NotFound: true, Type: RESOURCE_TYPE_GROUP, Name: name}
+}
+
+func AddGroupToUser(userEmail string, config GroupConfig, secret *coreV1.Secret) error {
+	token, err := GetTokenAsAdmin(secret)
+	if err != nil {
+		return err
+	}
+
+	userId, err := GetUserIdByEmail(userEmail, secret)
+	if err != nil {
+		return err
+	}
+
+	groupId, err := GetGroupIdByName(config.Name, secret)
+	if err != nil {
+		return err
+	}
+
+	params := map[string]string{
+		"userId":  userId,
+		"groupId": groupId,
+	}
+	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_ADD_GROUP_TO_USER, params)
+	req, err := http.NewRequest(http.MethodPut, url, nil)
+	if err != nil {
+		return err
+	}
+	req.Header.Add("Authorization", token)
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	if !IsOK(resp.StatusCode) {
+		reqDump, _ := httputil.DumpRequest(req, true)
+		return fmt.Errorf("failed to add group to user: " + string(reqDump) + "\n" + resp.Status)
+	}
+
+	return nil
+}
+
 func DeleteClient(config ClientConfig, secret *coreV1.Secret) error {
 	token, err := GetTokenAsAdmin(secret)
 	if err != nil {
@@ -589,11 +684,10 @@ func DeleteClient(config ClientConfig, secret *coreV1.Secret) error {
 	}
 
 	id, err := GetIdByClientId(config.ClientId, secret)
-	if err != nil {
-		return err
-	}
-	if !IsClientExist(id) {
+	if IsNotFound(err) {
 		return nil
+	} else if err != nil {
+		return err
 	}
 
 	params := map[string]string{
@@ -604,7 +698,45 @@ func DeleteClient(config ClientConfig, secret *coreV1.Secret) error {
 	if err != nil {
 		return err
 	}
-	req.Header.Add("Authorization", "Bearer "+token)
+	req.Header.Add("Authorization", token)
+
+	client := &http.Client{}
+	resp, err := client.Do(req)
+	if err != nil {
+		return err
+	}
+	defer resp.Body.Close()
+
+	if !IsOK(resp.StatusCode) {
+		reqDump, _ := httputil.DumpRequest(req, true)
+		return fmt.Errorf("failed to delete client: " + string(reqDump) + "\n" + resp.Status)
+	}
+
+	return nil
+}
+
+func DeleteGroup(config GroupConfig, secret *coreV1.Secret) error {
+	token, err := GetTokenAsAdmin(secret)
+	if err != nil {
+		return err
+	}
+
+	groupId, err := GetGroupIdByName(config.Name, secret)
+	if IsNotFound(err) {
+		return nil
+	} else if err != nil {
+		return err
+	}
+
+	params := map[string]string{
+		"groupId": groupId,
+	}
+	url := SetServiceDomainURI(KEYCLOAK_ADMIN_SERVICE_DELETE_GROUP, params)
+	req, err := http.NewRequest(http.MethodDelete, url, nil)
+	if err != nil {
+		return err
+	}
+	req.Header.Add("Authorization", token)
 
 	client := &http.Client{}
 	resp, err := client.Do(req)